new registration and login scenario using API

+++ app/src/main/java/com/matryoshka/projectx/utils/DateTimeUtils.kt
@@ -2,9 +2,9 @@ package com.matryoshka.projectx.utils
 
 import java.time.Duration
 
-//standard functions Duration.toHoursPart() and Duration.toMinutesPart() require API 31
-val Duration.hoursPart: Long
-    get() = this.toHours() % 24
+//standard functions Duration.to..Part() require API 31
+val Duration.hoursPart: Int
+    get() = (this.toHours() % 24).toInt()
 
-val Duration.minutesPart: Long
-    get() = this.toMinutes() % 60
+val Duration.minutesPart: Int
+    get() = (this.toMinutes() % 60).toInt()
